// src/StreamMaster.UI/lib/api.d.ts
export interface components {
	schemas: {
// ...
		CancelChannelRequest: {
			/** Format: int32 */
			smChannelId?: number;
		};
		CancelClientRequest: {
			uniqueRequestId?: string;
		};
		MoveToNextStreamRequest: {
			/** Format: int32 */
			smChannelId?: number;
		};
// ...
		ChannelMetric: {
			smStreamInfo: components["schemas"]["SMStreamInfo"] | null;
			clientStreams: components["schemas"]["ClientStreamsDto"][];
			metrics: components["schemas"]["StreamHandlerMetrics"];
			isFailed: boolean;
			name: string;
			id: string;
			sourceName: string;
			channelLogo: string | null;
			streamLogo: string | null;
			videoInfo: string | null;
		};
		ClientStreamsDto: {
			metrics: components["schemas"]["StreamHandlerMetrics"] | null;
			smStreamId: string;
			name: string;
			streamLogo: string | null;
			channelLogo: string | null;
			clientIPAddress: string | null;
			clientUserAgent: string | null;
			smChannelId: string;
		};
// ...
		CustomPlayList: {
			name?: string;
			logo?: string;
			folderNfo?: components["schemas"]["Movie"] | null;
			customStreamNfos?: components["schemas"]["CustomStreamNfo"][];
		};
// ...
	};
	responses: never;
	parameters: never;
	requestBodies: never;
	headers: never;
	pathItems: never;
}